Make ChunkArea under 32 Bytes [Improve CacheLine Alignment]
Revert "Make ChunkArea under 32 Bytes [Improve CacheLine Alignment]"

This reverts commit b34ad8a68f2be64a5419a648b20200b8adc23a89.

Revert "Make ChunkArea under 32 Bytes [Improve CacheLine Alignment]"

This reverts commit 1d6073622b12b03b39783e4ab3ab08acfebb82d2.

Update DrawBuffers.java

+++ src/main/java/net/vulkanmod/render/chunk/DrawBuffers.java
@@ -22,7 +22,7 @@ public class DrawBuffers {
 
     private static final int VERTEX_SIZE = PipelineManager.TERRAIN_VERTEX_FORMAT.getVertexSize();
     private static final int INDEX_SIZE = Short.BYTES;
-    public final int index;
+    public final short index;
     private final Vector3i origin;
     private final int minHeight;
 
@@ -34,7 +34,7 @@ public class DrawBuffers {
 
     public DrawBuffers(int index, Vector3i origin, int minHeight) {
 
-        this.index = index;
+        this.index = (short) index;
         this.origin = origin;
         this.minHeight = minHeight;
     }
